Struct hyper_rustls::HttpsConnectorBuilder

pub struct HttpsConnectorBuilder<State>(/* private fields */);

A builder for an HttpsConnector

This makes configuration flexible and explicit and ensures connector features match crate features

Examples

use hyper_rustls::HttpsConnectorBuilder;

    let https = HttpsConnectorBuilder::new()
    .with_webpki_roots()
    .https_only()
    .enable_http1()
    .build();

Implementations

impl ConnectorBuilder<WantsTlsConfig>

pub fn new() -> Self

Creates a new ConnectorBuilder

pub fn with_tls_config( self, config: ClientConfig, ) -> ConnectorBuilder<WantsSchemes>

Passes a rustls ClientConfig to configure the TLS connection

The alpn_protocols field is required to be empty (or the function will panic) and will be rewritten to match the enabled schemes (see enable_http1, enable_http2) before the connector is built.

pub fn with_native_roots(self) -> Result<ConnectorBuilder<WantsSchemes>>

Shorthand for using rustlsโ default crypto provider and safe defaults, with native roots.

See ConfigBuilderExt::with_native_roots

pub fn with_provider_and_native_roots( self, provider: impl Into<Arc<CryptoProvider>>, ) -> Result<ConnectorBuilder<WantsSchemes>>

Shorthand for using a custom CryptoProvider and native roots

See ConfigBuilderExt::with_native_roots

impl ConnectorBuilder<WantsSchemes>

pub fn https_only(self) -> ConnectorBuilder<WantsProtocols1>

Enforce the use of HTTPS when connecting

Only URLs using the HTTPS scheme will be connectable.

pub fn https_or_http(self) -> ConnectorBuilder<WantsProtocols1>

Allow both HTTPS and HTTP when connecting

HTTPS URLs will be handled through rustls, HTTP URLs will be handled by the lower-level connector.

impl ConnectorBuilder<WantsProtocols1>

pub fn enable_http1(self) -> ConnectorBuilder<WantsProtocols2>

Enable HTTP1

This needs to be called explicitly, no protocol is enabled by default

pub fn with_server_name_resolver( self, resolver: impl ResolveServerName + 'static + Sync + Send, ) -> Self

Override server name for the TLS stack

By default, for each connection hyper-rustls will extract host portion of the destination URL and verify that server certificate contains this value.

If this method is called, hyper-rustls will instead use this resolver to compute the value used to verify the server certificate.

pub fn with_server_name(self, override_server_name: String) -> Self

๐Deprecated since 0.27.1: use Self::with_server_name_resolver with FixedServerNameResolver instead

Override server name for the TLS stack

By default, for each connection hyper-rustls will extract host portion of the destination URL and verify that server certificate contains this value.

If this method is called, hyper-rustls will instead verify that server certificate contains override_server_name. Domain name included in the URL will not affect certificate validation.

impl ConnectorBuilder<WantsProtocols2>

pub fn build(self) -> HttpsConnector<HttpConnector>

This builds an HttpsConnector built on hyperโs default HttpConnector

pub fn wrap_connector<H>(self, conn: H) -> HttpsConnector<H>

This wraps an arbitrary low-level connector into an HttpsConnector