hyper_rustls

Module connector

Modules

builder 🔒

Structs

DefaultServerNameResolver

The default server name resolver, which uses the hostname in the URI.

FixedServerNameResolver

A server name resolver which always returns the same fixed name.

HttpsConnector

A Connector for the https scheme.

Traits

ResolveServerName

A trait implemented by types that can resolve a ServerName for a request.

Type Aliases

BoxError 🔒